Address race in query pack rebundling
When rebundling a query pack we move the pack to a temporary storage to
remove its dependencies and cache prior to recompiling.

Part of this recompilation is resolving the query pack's dependencies.
When done in parallel for multiple packs we can end up in a situation
where a pack dependency resolution is reading throught the packs finding
the correct dependency while it is being moved (because it a query pack
being rebundled). This results in IO errors where the CLI errors out
with `No content to map due to end-of-input`.

By copying and overwriting we should remediate this issue and allow the
dependency resolution to work with stale data (which isn't a problem in
this case because nothing is changed in the `qlpack.yml`).