Add option to specify the host to download the bundle from.
This is primarily to allow Actions run on GitHub Enterprise Server to be
able to fetch the bundle from GitHub.com instead of the local instance.
This can be used in cases where the GHES instance either doesn't have
github/codeql-action or the repo contains only older releases.

The Actions octokit instance does not allow configuration of the host,
so we need to use a plain octokit instance in that case.